requirements.txt
streamlit==1.54.0
llama-cpp-python==0.3.16
chromadb==0.6.3
sentence-transformers==3.4.1
numpy<2.0.0